+++ src/verb/verb.rs
@@ -105,7 +105,7 @@ impl Verb {
         }
         let (needs_selection, needs_another_panel) = match &execution {
             VerbExecution::Internal(ie) => (ie.needs_selection(), false),
-            VerbExecution::External(ee) => (
+            VerbExecution::External(ee) | VerbExecution::ShellCommand(ee) => (
                 ee.exec_pattern.has_selection_group(),
                 ee.exec_pattern.has_other_panel_group(),
             ),
@@ -284,7 +284,9 @@ impl Verb {
             // We can't determine before execution what will be the arguments, except
             // for the first item of the sequence. It's cleaner to just not try expand it
             format!("Hit *enter* to **{}**: `{}`", name, seq_ex.sequence.raw)
-        } else if let VerbExecution::External(external_exec) = &self.execution {
+        } else if let VerbExecution::External(external_exec)
+        | VerbExecution::ShellCommand(external_exec) = &self.execution
+        {
             let exec_desc = builder().shell_exec_string(&external_exec.exec_pattern, con);
             format!("Hit *enter* to **{}**: `{}`", name, exec_desc)
         } else if self.description.code {